Move 'gwb_ideal' raw calculation to gravwaves.py, and merge code in sam.py and simple_sam.py
