+++ holodeck/simple_sam.py
@@ -2,8 +2,8 @@
 """
 
 import numpy as np
-from holodeck import cosmo, utils
-from holodeck.constants import MSOL, GYR, MPC, SPLC, NWTG
+from holodeck import cosmo, utils, gravwaves
+from holodeck.constants import MSOL, GYR, MPC
 
 _AGE_UNIVERSE_GYR = cosmo.age(0.0).to('Gyr').value  # [Gyr]  ~ 13.78
 
@@ -141,20 +141,12 @@ class Simple_SAM:
         return rv
 
     def _zprime(self, mgal, qgal, redz):
-        print(f"{self}._zprime")
         tau0 = self.gmt(mgal, qgal, redz)  # sec
-        print(f"{utils.stats(tau0/GYR)=}")
         age = cosmo.age(redz).to('s').value
-        print(f"{utils.stats(age/GYR)=}")
         new_age = age + tau0
-        print(f"{utils.stats(new_age/GYR)=}")
-
         redz_prime = -1.0 * np.ones_like(new_age)
         idx = (new_age < _AGE_UNIVERSE_GYR * GYR)
         redz_prime[idx] = cosmo.tage_to_z(new_age[idx])
-        print(f"{utils.stats(redz)=}")
-        print(f"{utils.stats(redz_prime)=}")
-        print(f"{redz.shape=}, {new_age.shape=}, {redz_prime.shape=}")
         return redz_prime
 
     def gwb_sam(self, fobs_gw, sam, dlog10=True, sum=True):
@@ -172,26 +164,25 @@ class Simple_SAM:
         if not dlog10:
             ndens = ndens / (np.log(10.0) * mtot)
 
-        gwb = gwb_ideal(fobs_gw, ndens, mtot, mrat, rz, dlog10=dlog10, sum=sum)
-        # gwb = gwb_ideal_dlog10m(fobs_gw, ndens_dlog10m, mtot, mrat, rz)
+        gwb = gravwaves.gwb_ideal(fobs_gw, ndens, mtot, mrat, rz, dlog10=dlog10, sum=sum)
 
         return gwb
 
-    def gwb_ideal(self, fobs_gw, dlog10=True, sum=True):
+    def gwb_ideal(self, fobs_gw, dlog10=True, sum=True, redz_prime=True):
         # NOTE: dlog10M performs MUCH better than dM
         mg = self.mass_gal[:, np.newaxis, np.newaxis]
         qg = self.mrat_gal[np.newaxis, :, np.newaxis]
-        rz = self.redz[np.newaxis, np.newaxis, :]
-        # if redz_prime:
+        redz = self.redz[np.newaxis, np.newaxis, :]
+        ndens = self.ndens_mbh(mg, qg, redz, dlog10=dlog10) / (MPC**3)
+
         # convert from initial galaxy-merger redshift, to after galaxy merger-time
-        rz = self._zprime(mg, qg, rz)
+        if redz_prime:
+            redz = self._zprime(mg, qg, redz)
+            print(f"{self} :: {utils.stats(redz)=}")
 
         mtot = self.mbh[:, :, np.newaxis]
         mrat = self.qbh[np.newaxis, :, np.newaxis]
-
-        ndens = self.ndens_mbh(mg, qg, rz, dlog10=dlog10) / (MPC**3)
-
-        gwb = gwb_ideal(fobs_gw, ndens, mtot, mrat, rz, dlog10=dlog10, sum=sum)
+        gwb = gravwaves.gwb_ideal(fobs_gw, ndens, mtot, mrat, redz, dlog10=dlog10, sum=sum)
         return gwb
 
     def _integrated_ndens_mbh(self):
@@ -320,50 +311,3 @@ class Simple_SAM:
     def qgal_to_qbh(self, qgal):
         qbh = np.power(qgal, self._alpha_mbh_star)
         return qbh
-
-
-def gwb_ideal(fobs_gw, ndens, mtot, mrat, redz, dlog10, sum=True):
-
-    const = ((4.0 * np.pi) / (3 * SPLC**2))
-    mc = utils.chirp_mass_mtmr(mtot, mrat)
-    mc = np.power(NWTG * mc, 5.0/3.0)
-    rz = np.power(1 + redz, -1.0/3.0)
-    fogw = np.power(np.pi * fobs_gw, -4.0/3.0)
-
-    integ = ndens * mc * rz
-    redz = redz * np.ones_like(integ)
-    integ[redz <= 0.0] = 0.0
-
-    arguments = [mtot, mrat, redz]
-    if dlog10:
-        arguments[0] = np.log10(arguments[0])
-
-    for ax, xx in enumerate(arguments):
-        integ = np.moveaxis(integ, ax, 0)
-        xx = np.moveaxis(xx, ax, 0)
-
-        # if integ is (X, A, B) and xx is (X, 1, 1), then this is fine
-        try:
-            integ = 0.5 * (integ[:-1] + integ[1:]) * np.diff(xx, axis=0)
-        # BUT if integ is (X, A, B) and xx is (X, A+1, B+1), then need to average xx values down
-        except ValueError:
-            # average other dimensions as needed
-            for jj in range(1, len(arguments)):
-                sh = np.shape(xx)[jj]
-                if (sh == 1) or (sh == np.shape(integ)[jj]):
-                    continue
-
-                xx = np.moveaxis(xx, jj, 0)
-                xx = 0.5 * (xx[:-1] + xx[1:])
-                xx = np.moveaxis(xx, 0, jj)
-
-            # try integration step again
-            integ = 0.5 * (integ[:-1] + integ[1:]) * np.diff(xx, axis=0)
-
-        # return integ to the correct shape (axis order)
-        integ = np.moveaxis(integ, 0, ax)
-
-    gwb = const * fogw
-    gwb = gwb * np.sum(integ) if sum else gwb * integ
-    gwb = np.sqrt(gwb)
-    return gwb
